Fix: Set default accommodation preference
Set the default accommodation preference to "single" on the second step of the form.

+++ src/components/PriceQuoteForm.tsx
@@ -33,7 +33,7 @@ const PriceQuoteForm = ({ destinationName, availableDistances }: PriceQuoteFormP
     defaultValues: {
       preferredLanguage: 'english',
       preferredDistance: availableDistances[0],
-      accommodationPreference: 'double'
+      accommodationPreference: 'single' // Changed from 'double' to 'single'
     }
   });
